cleanup sampling code SampleFn -> GenSampleFn
