+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/core/RiskTestingFn.kt
@@ -30,5 +30,4 @@ interface RiskTestingFn {
         showDetails: Boolean = false,
         drawSample: () -> Double,
     ): TestH0Result
-
 }
