+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/core/Samples.kt
@@ -4,7 +4,7 @@ import org.cryptobiotic.rlauxe.util.doubleIsClose
 import kotlin.math.ln
 import kotlin.random.Random
 
-//// keeps track of the latest sample, number of samples, and the sample sum.
+/** keeps track of the latest sample, number of samples, and the sample sum. */
 interface Samples {
     fun last(): Double  // latest sample
     fun numberOfSamples(): Int    // total number of samples so far
@@ -16,8 +16,8 @@ interface Samples {
 /**
  * This ensures that the called function doesnt have access to the current sample,
  * as required by "predictable function of the data X1 , . . . , Xi−1" requirement.
- * Its up to the method using this to make it "previous samples", by not adding the current sample to it until
- * the end of the iteration.
+ * Its up to the method using this to make it "previous samples", by not adding the
+ * current sample to it until the end of the iteration.
  */
 class PrevSamples : Samples {
     private var last = 0.0
@@ -37,6 +37,7 @@ class PrevSamples : Samples {
     }
 }
 
+/** This also counts the under/overstatements. */
 class PrevSamplesWithRates(val noerror: Double) : Samples {
     private var last = 0.0
     private var sum = 0.0
@@ -65,7 +66,7 @@ class PrevSamplesWithRates(val noerror: Double) : Samples {
         sum += sample
         welford.update(sample)
 
-        // or just say which overstatement it is
+        // or just say which overstatement it is?
         if (doubleIsClose(sample, noerror)) countP0++
         if (doubleIsClose(sample, noerror * 0.5)) countP1++
         if (doubleIsClose(sample, 0.0)) countP2++
@@ -74,379 +75,3 @@ class PrevSamplesWithRates(val noerror: Double) : Samples {
     }
 }
 
-//// abstraction for creating a sequence of samples
-interface SampleFn { // TODO could be an Iterator
-    fun sample(): Double // get next in sample
-    fun reset()          // start over again with different permutation
-    fun sampleMean(): Double // for simulations
-    fun sampleCount(): Double
-    fun N(): Int  // population size
-}
-
-///////////////////////////////////////////////////////////////
-// For polling audits.
-
-class PollWithReplacement(val cvrs : List<Cvr>, val assorter: AssorterFunction): SampleFn {
-    val N = cvrs.size
-    val sampleMean = cvrs.map { assorter.assort(it) }.average()
-    val sampleCount = cvrs.sumOf { assorter.assort(it) }
-
-    override fun sample(): Double {
-        val idx = Random.nextInt(N) // with Replacement
-        return assorter.assort(cvrs[idx])
-    }
-
-    override fun reset() {}
-    override fun sampleMean() = sampleMean
-    override fun sampleCount() = sampleCount
-    override fun N() = N
-}
-
-class PollWithoutReplacement(val cvrs : List<Cvr>, val assorter: AssorterFunction): SampleFn {
-    val N = cvrs.size
-    val permutedIndex = MutableList(N) { it }
-    var idx = 0
-
-    init {
-        reset()
-    }
-
-    override fun sample(): Double {
-        require (idx < cvrs.size)
-        require (permutedIndex[idx] < cvrs.size)
-        val curr = cvrs[permutedIndex[idx++]]
-        return assorter.assort(curr)
-    }
-
-    override fun reset() {
-        permutedIndex.shuffle(Random)
-        idx = 0
-    }
-
-    override fun sampleMean() = cvrs.map{ assorter.assort(it) }.average()
-    override fun sampleCount() = cvrs.sumOf { assorter.assort(it) }
-    override fun N() = N
-}
-
-///////////////////////////////////////////////////////////////
-// the values produced here are the B assort values, SHANGRLA section 3.2.
-
-// the mvr and cvr always agree.
-class ComparisonNoErrors(val cvrs : List<Cvr>, val cassorter: ComparisonAssorter): SampleFn {
-    val N = cvrs.size
-    val permutedIndex = MutableList(N) { it }
-    val sampleMean: Double
-    val sampleCount: Double
-    var idx = 0
-
-    init {
-        reset()
-        sampleMean = cvrs.map { cassorter.bassort(it, it)}.average()
-        sampleCount = cvrs.sumOf { cassorter.bassort(it, it) }
-    }
-
-    override fun sample(): Double {
-        require (idx < N)
-        val curr = cvrs[permutedIndex[idx++]]
-        return cassorter.bassort(curr, curr) // mvr == cvr, no errors. could just return cassorter.noerror
-    }
-
-    override fun reset() {
-        permutedIndex.shuffle(Random)
-        idx = 0
-    }
-
-    override fun sampleMean() = sampleMean
-    override fun sampleCount() = sampleCount
-    override fun N() = N
-}
-
-// generate mvr by starting with cvrs and flipping exact # votes (type 2 errors only) to make mvrs have mvrMean
-// withReplacement
-data class ComparisonWithErrors(val cvrs : List<Cvr>, val cassorter: ComparisonAssorter, val mvrMean: Double): SampleFn {
-    val N = cvrs.size
-    val mvrs : List<Cvr>
-    val permutedIndex = MutableList(N) { it }
-    val sampleMean: Double
-    val sampleCount: Double
-    val flippedVotes: Int
-    var idx = 0
-
-    init {
-        reset()
-
-        // we want to flip the exact number of votes, for reproducibility
-        val mmvrs = mutableListOf<Cvr>()
-        mmvrs.addAll(cvrs)
-        flippedVotes = flipExactVotes(mmvrs, mvrMean)
-        mvrs = mmvrs.toList()
-
-        sampleCount = cvrs.mapIndexed { idx, it -> cassorter.bassort(mvrs[idx], it)}.sum()
-        sampleMean = sampleCount / N
-    }
-
-    override fun sample(): Double {
-        val cvr = cvrs[permutedIndex[idx]]
-        val mvr = mvrs[permutedIndex[idx]]
-        idx++
-        return cassorter.bassort(mvr, cvr)
-    }
-
-    override fun reset() {
-        permutedIndex.shuffle(Random)
-        idx = 0
-    }
-
-    override fun sampleMean() = sampleMean
-    override fun sampleCount() = sampleCount
-    override fun N() = N
-}
-
-// generate mvr by starting with cvrs and flipping (N * p2) votes (type 2 errors) and (N * p1) votes (type 1 errors)
-// withReplacement
-data class ComparisonWithErrorRates(val cvrs : List<Cvr>, val cassorter: ComparisonAssorter,
-                                    val p2: Double, val p1: Double = 0.0, val withoutReplacement: Boolean = true): SampleFn {
-    val N = cvrs.size
-    val mvrs : List<Cvr>
-    val permutedIndex = MutableList(N) { it }
-    val sampleMean: Double
-    val sampleCount: Double
-    val flippedVotes2: Int
-    val flippedVotes1: Int
-
-    var idx = 0
-
-    init {
-        reset()
-
-        // we want to flip the exact number of votes, for reproducibility
-        val mmvrs = mutableListOf<Cvr>()
-        mmvrs.addAll(cvrs)
-        flippedVotes2 = add2voteOverstatements(mmvrs, needToChangeVotesFromA = (N * p2).toInt())
-        flippedVotes1 =  if (p1 == 0.0) 0 else {
-            add1voteOverstatements(mmvrs, needToChangeVotesFromA = (N * p1).toInt())
-        }
-        mvrs = mmvrs.toList()
-
-        sampleCount = cvrs.mapIndexed { idx, it -> cassorter.bassort(mvrs[idx], it)}.sum()
-        sampleMean = sampleCount / N
-    }
-
-    override fun sample(): Double {
-        val assortVal = if (withoutReplacement) {
-            val cvr = cvrs[permutedIndex[idx]]
-            val mvr = mvrs[permutedIndex[idx]]
-            idx++
-            cassorter.bassort(mvr, cvr)
-        } else {
-            val chooseIdx = Random.nextInt(N) // with Replacement
-            val cvr = cvrs[chooseIdx]
-            val mvr = mvrs[chooseIdx]
-            cassorter.bassort(mvr, cvr)
-        }
-        return assortVal
-    }
-
-    override fun reset() {
-        permutedIndex.shuffle(Random)
-        idx = 0
-    }
-
-    override fun sampleMean() = sampleMean
-    override fun sampleCount() = sampleCount
-    override fun N() = N
-}
-
-///////////////////////
-
-// change cvrs to have the exact number of votes for wantAvg
-fun flipExactVotes(cvrs: MutableList<Cvr>, wantAvg: Double): Int {
-    val ncards = cvrs.size
-    val expectedAVotes = (ncards * wantAvg).toInt()
-    val actualAvotes = cvrs.sumOf { it.hasMarkFor(0, 0) }
-    val needToChangeVotesFromA = actualAvotes - expectedAVotes
-    return add2voteOverstatements(cvrs, needToChangeVotesFromA)
-}
-
-// change cvrs to add the given number of two-vote overstatements.
-// Note that we replace the Cvr in the list when we change it
-fun add2voteOverstatements(cvrs: MutableList<Cvr>, needToChangeVotesFromA: Int): Int {
-    val ncards = cvrs.size
-    val startingAvotes = cvrs.sumOf { it.hasMarkFor(0, 0) }
-    var changed = 0
-    // we need more A votes, needToChangeVotesFromA < 0>
-    if (needToChangeVotesFromA < 0) {
-        while (changed > needToChangeVotesFromA) {
-            val cvrIdx = Random.nextInt(ncards)
-            val cvr = cvrs[cvrIdx]
-            if (cvr.hasMarkFor(0, 1) == 1) {
-                val votes = mutableMapOf<Int, Map<Int, Int>>()
-                votes[0] = mapOf(0 to 1)
-                cvrs[cvrIdx] = Cvr("card-$cvrIdx", votes)
-                changed--
-            }
-        }
-    } else {
-        // we need more B votes, needToChangeVotesFromA > 0
-        while (changed < needToChangeVotesFromA) {
-            val cvrIdx = Random.nextInt(ncards)
-            val cvr = cvrs[cvrIdx]
-            if (cvr.hasMarkFor(0, 0) == 1) {
-                val votes = mutableMapOf<Int, Map<Int, Int>>()
-                votes[0] = mapOf(1 to 1)
-                cvrs[cvrIdx] = Cvr("card-$cvrIdx", votes)
-                changed++
-            }
-        }
-    }
-    val checkAvotes = cvrs.sumOf { it.hasMarkFor(0, 0) }
-    // if (debug) println("flipped = $needToChangeVotesFromA had $startingAvotes now have $checkAvotes votes for A")
-    require(checkAvotes == startingAvotes - needToChangeVotesFromA)
-    return changed
-}
-
-// change cvrs to add the given number of one-vote overstatements.
-fun add1voteOverstatements(cvrs: MutableList<Cvr>, needToChangeVotesFromA: Int): Int {
-    val ncards = cvrs.size
-    val startingAvotes = cvrs.sumOf { it.hasMarkFor(0, 0) }
-    var changed = 0
-    while (changed < needToChangeVotesFromA) {
-        val cvrIdx = Random.nextInt(ncards)
-        val cvr = cvrs[cvrIdx]
-        if (cvr.hasMarkFor(0, 0) == 1) {
-            val votes = mutableMapOf<Int, Map<Int, Int>>()
-            votes[0] = mapOf(2 to 1)
-            cvrs[cvrIdx] = Cvr("card-$cvrIdx", votes)
-            changed++
-        }
-    }
-    val checkAvotes = cvrs.sumOf { it.hasMarkFor(0, 0) }
-    // if (debug) println("flipped = $needToChangeVotesFromA had $startingAvotes now have $checkAvotes votes for A")
-    require(checkAvotes == startingAvotes - needToChangeVotesFromA)
-    return changed
-}
-
-
-///////////////////////
-//// DoubleArrays
-fun randomPermute(samples : DoubleArray): DoubleArray {
-    val n = samples.size
-    val permutedIndex = MutableList(n) { it }
-    permutedIndex.shuffle(Random)
-    return DoubleArray(n) { samples[permutedIndex[it]] }
-}
-
-// generate a sample thats approximately mean = theta
-fun generateUniformSample(N: Int) : DoubleArray {
-    return DoubleArray(N) {
-        Random.nextDouble(1.0)
-    }
-}
-
-// generate a sample thats approximately mean = theta
-fun generateSampleWithMean(N: Int, ratio: Double) : DoubleArray {
-    return DoubleArray(N) {
-        val r = Random.nextDouble(1.0)
-        if (r < ratio) 1.0 else 0.0
-    }
-}
-
-class ArrayAsSampleFn(val assortValues : DoubleArray): SampleFn {
-    var index = 0
-
-    override fun sample(): Double {
-        return assortValues[index++]
-    }
-
-    override fun reset() {
-        index = 0
-    }
-
-    override fun sampleMean(): Double {
-        return assortValues.toList().average()
-    }
-
-    override fun sampleCount(): Double {
-        return assortValues.toList().sum()
-    }
-
-    override fun N(): Int {
-        return assortValues.size
-    }
-}
-
-// generate random values with given mean
-class SampleMeanWithReplacement(val N: Int, ratio: Double): SampleFn {
-    val samples = generateSampleWithMean(N, ratio)
-    override fun sample(): Double {
-        val idx = Random.nextInt(N) // with Replacement
-        return samples[idx]
-    }
-    override fun reset() {
-        // noop
-    }
-    override fun sampleMean() = samples.average()
-    override fun sampleCount() = samples.sum()
-    override fun N() = N
-}
-
-class SampleMeanWithoutReplacement(val N: Int, val ratio: Double): SampleFn {
-    var samples = generateSampleWithMean(N, ratio)
-    var index = 0
-    override fun sample(): Double {
-        return samples[index++]
-    }
-    override fun reset() {
-        samples = generateSampleWithMean(N, ratio)
-        index = 0
-    }
-    override fun sampleMean() = samples.average()
-    override fun sampleCount() = samples.sum()
-    override fun N() = N
-}
-
-class SampleFromArrayWithoutReplacement(val assortValues : DoubleArray): SampleFn {
-    val N = assortValues.size
-    val permutedIndex = MutableList(N) { it }
-    var idx = 0
-
-    init {
-        reset()
-    }
-
-    override fun sample(): Double {
-        require (idx < N)
-        require (permutedIndex[idx] < N)
-        return assortValues[permutedIndex[idx++]]
-    }
-
-    override fun reset() {
-        permutedIndex.shuffle(Random)
-        idx = 0
-    }
-
-    override fun sampleCount() = assortValues.sum()
-    override fun sampleMean() = assortValues.average()
-    override fun N() = N
-}
-
-///////////////////////////////////////////////////////////////
-
-// generate Bernoulli with probability p
-class Bernoulli(p: Double) {
-    val log_q = ln(1.0 - p)
-    val n = 1.0
-
-    fun get(): Double {
-        var x = 0.0
-        var sum = 0.0
-        while (true) {
-            val wtf = ln( Math.random()) / (n - x)
-            sum += wtf
-            if (sum < log_q) {
-                return x
-            }
-            x++
-        }
-    }
-}
