+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/comparison/TestBravo.kt
@@ -55,7 +55,7 @@ class TestBravo  {
         trueMeans.forEach { trueMean ->
             val estimFn = FixedMean(eta0)
             val alpha = AlphaMart(estimFn = estimFn, N = N, upperBound = 1.0, withoutReplacement = withoutReplacement)
-            val sampler = if (withoutReplacement) SampleMeanWithoutReplacement(N, trueMean) else SampleMeanWithReplacement(N, trueMean)
+            val sampler = if (withoutReplacement) GenSampleMeanWithoutReplacement(N, trueMean) else GenSampleMeanWithReplacement(N, trueMean)
 
             val rr = runTestRepeated(
                 drawSample = sampler,
