+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/polling/PollingParameters.kt
@@ -1,7 +1,6 @@
 package org.cryptobiotic.rlauxe.polling
 
 import org.cryptobiotic.rlauxe.rlaplots.SRTcsvWriter
-import org.cryptobiotic.rlauxe.sim.PollingRunner
 import org.cryptobiotic.rlauxe.sim.PollingTask
 import org.cryptobiotic.rlauxe.sim.RepeatedTaskRunner
 import org.cryptobiotic.rlauxe.util.makeCvrsByExactMean
